nfs-utils:  Add structure for passing local binding info.
Create a struct local_bind_info that will hold an IP
address to be used for binding to a local IP address.
Pass this down to the code that needs it.

There are no functional changes in this patch, just passing
around un-used variables which are always NULL.  Future
patches will use this.

Signed-off-by: Ben Greear <greearb@candelatech.com>